% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsFineliFoodDataTables}
\alias{lsFineliFoodDataTables}
\title{Fineli (Food Data) Tables}
\format{
A list of list with named elements.
\describe{
  \item{names}{Language.}
}
}
\source{
\url{https://fineli.fi/fineli/en/avoin-data}
}
\usage{
lsFineliFoodDataTables
}
\description{
A list of data frames that contain food data
collected and organized by National Institute for Health and Welfare, Fineli.
}
\details{
The content of the Fineli database is freely available as open data.
The use of the material is governed by the terms and conditions specified by license: Creative Commons Attribution 4.0 (CC-BY 4.0).
When used, the material or the service including or
utilising the material must be accompanied by information on the original source and the date of the current version as follows:
National Institute for Health and Welfare, Fineli.
}
\keyword{datasets}
