% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineliFoodDataFunctions.R
\name{MakeDictionary}
\alias{MakeDictionary}
\title{Make dictionary}
\usage{
MakeDictionary(
  table,
  nameFrom = NULL,
  nameTo = NULL,
  lang = "English",
  columnNamePrefixesQ = FALSE,
  sep = ":"
)
}
\arguments{
\item{table}{A data frame or a string of Fineli Food Data table name.
(Generic names are allowed, without indication of language.)}

\item{nameFrom}{A string.
If NULL the first integer column is used.}

\item{nameTo}{A string.
If NULL the first character column is used.}

\item{lang}{Language for the interpretations.
One of \code{c("EN", "English", "FI", "Finnish", "SV", "Swedish")}.
If NULL then it is the same as "English".}

\item{columnNamePrefixesQ}{Should the column names be used as prefixes or not?}

\item{sep}{Separator for the prefixes of the columns names.}
}
\value{
A list with named elements.
}
\description{
Makes an (interpretation) dictionary
for a specified table, columns, and language.
}
