% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineliFoodDataFunctions.R
\name{SummarizeTables}
\alias{SummarizeTables}
\title{Summarize tables}
\usage{
SummarizeTables(tables = NULL)
}
\arguments{
\item{tables}{A list of tables (data frame).
If NULL then the Fineli Food Data list of tables \code{lsFineliFoodDataTables} is used.}
}
\value{
A data frame
}
\description{
Summarizes a list of tables.
}
